refactor: common CLI helpers
