+++ app/client/cli/debug.go
@@ -1,8 +1,6 @@
 package cli
 
 import (
-	"errors"
-	"fmt"
 	"os"
 
 	"github.com/manifoldco/promptui"
@@ -11,10 +9,7 @@ import (
 
 	"github.com/pokt-network/pocket/app/client/cli/helpers"
 	"github.com/pokt-network/pocket/logger"
-	"github.com/pokt-network/pocket/p2p/providers/peerstore_provider"
-	typesP2P "github.com/pokt-network/pocket/p2p/types"
 	"github.com/pokt-network/pocket/shared/messaging"
-	"github.com/pokt-network/pocket/shared/modules"
 )
 
 // TECHDEBT: Lowercase variables / constants that do not need to be exported.
@@ -184,7 +179,7 @@ func sendDebugMessage(cmd *cobra.Command, debugMsg *messaging.DebugMessage) {
 		logger.Global.Error().Err(err).Msg("Failed to create Any proto")
 	}
 
-	pstore, err := fetchPeerstore(cmd)
+	pstore, err := helpers.FetchPeerstore(cmd)
 	if err != nil {
 		logger.Global.Fatal().Err(err).Msg("Unable to retrieve the pstore")
 	}
@@ -203,46 +198,8 @@ func sendDebugMessage(cmd *cobra.Command, debugMsg *messaging.DebugMessage) {
 		logger.Global.Fatal().Err(err).Msg("Failed to convert validator address into pocketCrypto.Address")
 	}
 
+	// TECHDEBT: prefer to retrieve P2P module from the bus instead.
 	if err := helpers.P2PMod.Send(validatorAddress, anyProto); err != nil {
 		logger.Global.Error().Err(err).Msg("Failed to send debug message")
 	}
-}
-
-// fetchPeerstore retrieves the providers from the CLI context and uses them to retrieve the address book for the current height
-func fetchPeerstore(cmd *cobra.Command) (typesP2P.Peerstore, error) {
-	bus, ok := helpers.GetValueFromCLIContext[modules.Bus](cmd, helpers.BusCLICtxKey)
-	if !ok || bus == nil {
-		return nil, errors.New("retrieving bus from CLI context")
-	}
-	// TECHDEBT(#810, #811): use `bus.GetPeerstoreProvider()` after peerstore provider
-	// is retrievable as a proper submodule
-	pstoreProvider, err := bus.GetModulesRegistry().GetModule(peerstore_provider.PeerstoreProviderSubmoduleName)
-	if err != nil {
-		return nil, errors.New("retrieving peerstore provider")
-	}
-	currentHeightProvider := bus.GetCurrentHeightProvider()
-
-	height := currentHeightProvider.CurrentHeight()
-	pstore, err := pstoreProvider.(peerstore_provider.PeerstoreProvider).GetStakedPeerstoreAtHeight(height)
-	if err != nil {
-		return nil, fmt.Errorf("retrieving peerstore at height %d", height)
-	}
-	// Inform the client's main P2P that a the blockchain is at a new height so it can, if needed, update its view of the validator set
-	err = sendConsensusNewHeightEventToP2PModule(height, bus)
-	if err != nil {
-		return nil, errors.New("sending consensus new height event")
-	}
-	return pstore, nil
-}
-
-// sendConsensusNewHeightEventToP2PModule mimicks the consensus module sending a ConsensusNewHeightEvent to the p2p module
-// This is necessary because the debug client is not a validator and has no consensus module but it has to update the peerstore
-// depending on the changes in the validator set.
-// TODO(#613): Make the debug client mimic a full node.
-func sendConsensusNewHeightEventToP2PModule(height uint64, bus modules.Bus) error {
-	newHeightEvent, err := messaging.PackMessage(&messaging.ConsensusNewHeightEvent{Height: height})
-	if err != nil {
-		logger.Global.Fatal().Err(err).Msg("Failed to pack consensus new height event")
-	}
-	return bus.GetP2PModule().HandleEvent(newHeightEvent.Content)
 }
